requirements.txt
dvc==1.8.4
joblib==0.17.0
jupyter==1.0.0
jupyter_contrib_nbextensions==0.5.1
ipykernel==5.3.4
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
pyyaml==5.3.1
scikit-learn==0.23.2
scipy==1.5.3
tqdm==4.50.2